+++ actix-http/CHANGES.md
@@ -6,6 +6,9 @@
 
 * Bump minimum supported Rust version to 1.40
 
+* `BodySize::Sized64` variant has been removed. `BodySize::Sized` now receives a
+  `u64` instead of a `usize`.
+
 ### Fixed
 
 * Support parsing of `SameSite=None` [#1503]